requirements.txt
fastapi
uvicorn
pillow
imagehash
opencv-python-headless
python-multipart
google-genai
python-dotenv
requests
